Updated openstack/openstack Project: openstack-dev/devstack  a25ae6c03de2d87bca4f708ffc338c66164eec3b
convert apache logs to color for screen

If we left the ansi color codes in apache logs, we can run a sed
script to convert the escaped escapes back to ansi escapes which make
the logs colorized again.

There are 8 \ because we need to end up with 2 in the final sed, and
we get interopolated twice. How much fun is escape interpolation? All
the fun.

Change-Id: Id8531cf03ba80f0df62f20add02e757bd63d4f2d