Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  17bd56dda9ed9174325f4265a3d813ea39631472
Remove unused builder.Builder.update_job method

The deprecation warning message in this method was confusing and
seemed to be intended for builder.Jenkins.update_job but it's hard to
tell for sure. Seems likely to be an artifact of an outdated patchset
in the parallelize-the-things work done last year because it's not
clear how it could apply to builder.Builder.update_jobs which has a
significantly different use case than parallel_update_job.

Change-Id: Ifb3c1a40d81b0d43ac8e85151f0a99d1502e21db