Updated openstack/openstack Project: openstack/cinder  d2c3241115ce7bc1ae1997b23b5b22a0e943f663
Add group snapshots - manager

This is the fourth patch that implements the generic-volume-group
bluerpint. It implements group snapshots and create group from
source in the volume manager.

It depends on the third patch which adds db and objects for group
snapshots:
    https://review.openstack.org/#/c/328052/

The next patch adds group snapshots APIs:
    https://review.openstack.org/#/c/361369/

Change-Id: I5191a4cd1d82cc3598d71f2460d8239bdf1869c7
Partial-Implements: blueprint generic-volume-group