Updated openstack/openstack Project: openstack/neutron  3d1915ffe74c69a27ed566aef5ee3eab9a99ad30
Implement state management for trunks

This implements state management for trunks, allowing the trunk
state machine to be driven by the various events in the
lifecycle of a trunk. All events are handled here with the exception
of the unbinding of the parent port of a trunk.

Change-Id: I08341317f2cf6f4d1fc798773448ca0a79b242f5
Partially-implements: blueprint vlan-aware-vms