Updated openstack/openstack Project: openstack/neutron  92bcaed7aa63b49265f5f4fd4a9bc0a3ba95682d
Reduce DB hit when checking for trunk-able ports

We can hit the DB once to fetch all agent types for the
host rather than going back and forth for all the drivers
available in the deployment. That lets us save N-1 queries
where N is the number of drivers installed.

Change-Id: I5db0d73245b4e1c1bc46c063c0689ac3c5ea2bea