Updated openstack/openstack Project: openstack/horizon  9db0f0c0fdf32f21e22a3263332dcff29b42ccf2
Hard-code stubbed novaclient api_version to 2.1

The unit tests fail with novaclient 6.0.0 which requires
the use of an api_version property on the base Client class.

This change hard-codes the stubbed api_version attribute to
be 2.1 which is equivalent to the preferred version used in
the actual openstack_dashboard.api.nova.VERSSIONS code.

Change-Id: I9b89a224ad715cd0933d9757ca70e8a577eefc95
Closes-Bug: #1619093