Updated openstack/openstack Project: openstack/nova  92d64d8d071a330328d2bfe947c2bed5d60e1577
Eliminate additional DB queries in nova lists

Servers detail call has the extension which adds additional attribute
host_status to the instance (server). This extension mechanism issues a DB
query for each instance.host which can be an overhead for installation
with the large number of hosts. This change eliminates additional queries
and replaces them with a join query.

Before this change, it was K + N queries to DB
where K is a number of queries to get instances list and N is a number of
queries to get services for each instance.host.
After this change, it's only K queries.

Change-Id: Ia9331a2385b448c26ca7e281674a26f43e3f44de
Related-Bug: #1610153