Updated openstack/openstack Project: openstack/glance  cafea68016dd6837a5a8d550aa27b99fd02f7336
Release note for glance config opts.

Adding a release note stating the work on improving configuration
options of Glance for this release.
Some of the configuration options in Glance were set
with a ``min``, preventing them to accept negative integers like
in the previous release. This release note will give operators a
heads up on this.

This note is copied and modifed for glance from
I8db8764f84c92e987a774ce3650beb53d9635905
which was originally done for glance_store

Co-Authored-By: Dharini Chandrasekar <dharini.chandrasekar@intel.com>

Change-Id: I1afaa7206bd964409c343abcc385eff9dbf6b2ed