Updated openstack/openstack Project: openstack-infra/project-config  d9e751f8a9fb1bd22a7a6edb6df8df6218af5ddb
import the dependencies needed for creating stable branches

We have the script to create stable branches from the deliverable files,
but the thing it calls wasn't imported at the same time. This patch
moves the needed pieces from the release-tools repo to project-config so
they are all in the same place.

Change-Id: Ic4c3b68e1c98bda3dcbbc37ecc5f39ad5ac8f57f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>