Updated openstack/openstack Project: openstack-infra/project-config  7b3d934db8fa42fc9fe44186977e656d8b1c490a
Move unsuccessful non-voting OSA jobs to experimental

This patch moves many (but not all) of the experimental CentOS
jobs in OpenStack-Ansible from non-voting to experimental.

As this work will not be completed for Ocata it is pointless
for us to waste CI resources on executing jobs which will simply
fail.

Once we're ready to tackle them again we can complete the work,
test it experimentally, then graduate the job appropriately.

Change-Id: I95f0e01574e256ae79a8b66b169ee3e091528075