Updated openstack/openstack Project: openstack-infra/project-config  9ffa5c966d24f9bc7af171689e766198bea17179
be more careful using setuptools commands in release script

If the setup_requires dependencies are not installed yet, running
setuptools commands will install them. Capturing the output of a
setuptools command that includes the output from installing packages
produces a bad dist_name, so we first ask for the name without saving
the output and then we ask for it again and save the output to get a
clean version. This is why we can't have nice things.

Change-Id: I439d288566a4bd0cfd6f8a91d4136ed5a39d17cb
Signed-off-by: Doug Hellmann <doug@doughellmann.com>