Updated openstack/openstack Project: openstack/ironic  65da2b30c8379e79c1fbf2f1b4dd230adfc2a4a5
Fix updating port.portgroup_uuid for node

Port RPC object doesn't have portgroup_uuid attribute. As result it was
allowed to update portgroup_uuid when node is in ACTIVE/DEPLOYING and other
blocked for updates states.
This patch uses portgroup_id as it is a valid value for RPC object.

Change-Id: I3e849e78dfde9d21fc511cf92270aa7436a535ee