Updated openstack/openstack Project: openstack/ironic  b7b38352ea6c8d0004041f9a51986c628e67fdf9
DevStack: Configure retrieving logs from the deploy ramdisk

This patch changes the devstack plugin to allow users to configure the
ironic service to retrieve (or not) the logs from the deploy ramdisk. By
default this feature is enabled in Ironic to collect the logs upon a
deployment failure but in DevStack we've set it to *always* to help us
with troubleshooting.

Three new configuration were added to devstack:

1. $IRONIC_DEPLOY_LOGS_COLLECT - When Ironic should collect the logs,
   possible values are:

   * on_failure: Collect the ramdisk logs upon a deployment failure
   * always: Collect the ramdisk logs from the deployment on success or
     failure
   * never: Never collect the ramdisk logs

2. $IRONIC_DEPLOY_LOGS_STORAGE_BACKEND - Where should Ironic store the
   logs, possible values are:

   * local: To store the logs in the local filesystem
   * swift: To store the logs in Swift

3. $IRONIC_DEPLOY_LOGS_LOCAL_PATH - The path to the directory where
   Ironic should put the logs when $IRONIC_DEPLOY_LOGS_STORAGE_BACKEND is
   set to "local". Defaults to /opt/stack/ironic-bm-logs/deploy_logs

Change-Id: Iedd0858a84eff03d9b5074c04c9e132a90cea1cf
Closes-Bug: #1616969