Updated openstack/openstack Project: openstack-infra/storyboard  4b8ffd6bfed4b4d3d33058a471b02cd14e53ed5e
Limit PyMySQL to < 0.7.7

The tests are failing with weird foreign key constraint failures
on patches which don't affect the relevant bits of code.

This commit limits the version of PyMySQL to less than the recent
release (at the time of writing) which appears to correlate with
the test failures appearing.

Change-Id: I2924feed790786fb029f9cbbd89e186cb8f02967