Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  6853cf5ae7030fb82650a0ba5c6556e621f54313
Fix disabled always returning true

If disabled is configured to anything the existing code returns true.
Typically what was done to avoid this was to ensure disabled was
returned in the job as {obj:disabled} to keep the original type. However
if the user uses a JJB 'defaults', defaults always return a string so
{obj:disabled} does not receive a boolean it's expecting causing
disabled's value to always be a string which in code returns as true
always.

Change-Id: Ifd391945fbc7478a10d794e50a61559688c1a23e
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>