Updated openstack/openstack Project: openstack/nova  e9d548cb3497e58671612be1f005ef70518967bd
Run cell0 db migrations during nova-manage simple_cell_setup

In order for the cell0 database to be usable it needs to have db
migrations run against it. Since the simple_cell_setup attempts to be
the one command necessary to upgrade to cellsv2 it should also handle
the db migrations.

Partially-Implements: bp cells-cell0
Change-Id: I9bbaa4c92503222c9fd015fe075926b50f3dcc8c
Depends-On: I090d5994c3077d86c60457e4c16a03d35e0c409c