Updated openstack/openstack Project: openstack-dev/grenade  b69f4851950d2e6aa66617cf5b751fb8aa1c1324
Allow upgrade to create fernet keys for credentials

With the addition of encrypted credentials, keystone needs the ability to read
encryption keys from a repository in order to create credentials. This commit
makes it so that when keystone is upgraded from Mitaka to Newton, we run
`keystone-manage credential_setup` to ensure a directory exists on the system.

Depends-On: Ib06c86a17e585f43bfa8aedf3d5e89d3163cc9da
Change-Id: Id3e8922adc154cfec5f7a36613e22eb0b49eeffe
partially-implements: credential-encryption