Updated openstack/openstack Project: openstack/nova  62537a0618d38f90d12d5149aea69ff3fdee866c
Remove BuildRequest when scheduling fails

If there's a scheduler failure the instance is put into an error status
and that's what should be returned by an instance show/list. However
because the BuildRequest is not deleted that's what actually gets
returned. Delete the BuildRequest on scheduler failures.

Change-Id: Ia6c87aecc1484a2366d50822b669291c9f954545
Partially-implements: bp add-buildrequest-obj