Updated openstack/openstack Project: openstack/nova  bcaef0dee04200363825a1729b25b90580b355c6
Delete BuildRequest regardless of service_version

Because an instance list will attempt to pull instances from the
build_requests table we should delete the BuildRequest when attempting
to delete an instance no matter what the minimum service version is.
However the deletion only affects the rest of the delete logic if the
service version is at a certain level.

Change-Id: I0637be3b9357de0f0f8fa2e7bae00859ccbe18e1