Updated openstack/openstack Project: openstack/ceilometer  fe2dec6cf10305f9e3445c286ebb60d061aacb55
generalise instable API redirect assertion

these tests randomly fail because we check to verify tests explicitly
tried to grab URL from keystone and this fails because other stuff also
does this. pecan routing naturally inspects attributes and it's inspection
may trigger off unrelated url requests to be made.

this generalises the tests so we just ensure that we check keystone for url
regardless if it was done by us and regardless if other stuff called keystone
as well

Change-Id: Icbb4cd34f9c1aeb963fbd2d2d579e0f1cd9d0a60