Updated openstack/openstack Project: openstack/glance_store  b5833a87236183444589f3f93fba8641eb6c520a
Always return a sorted list of drivers for configs

glance.store entry point used in Glance for the store library is
resulting in generating random order in each run of the ``tox -e
genconfig`` command. This is because of the unsorted iterator returned
by the stevedore extension manager [1, 2].

This commit fixes that and returs a deterministic, sorted list of
drivers to generate the configs in Glance.

Example scenario is at change Iea2c538b37182191445cc5c1f8b2d9dceed06343

Closes-Bug: 1619487

[1] https://github.com/openstack/glance_store/blob/22fa2b08e76a8ebc39ade92504aeafb0bf8d2f8c/glance_store/backend.py#L135
[2] http://docs.openstack.org/developer/stevedore/managers.html#stevedore.extension.ExtensionManager.__iter__

Co-Authored-By: Joshua Harlow <jxharlow@godaddy.com>

Change-Id: I3146b8597a5d89da49b84d6653edacc3067c2c71