Updated openstack/openstack Project: openstack/nova  0c5e0f8ebbebe55b1b3383ab0c0f63f05ea986b5
Add create_all and delete_all for AllocationList

This is the object side of creating and deleting allocations in the
placement service.

create_all() checks that the inventory records involved in the
allocations requested do not have their capacities exceeded. If no
inventory capacities are exceeded, allocation records are written to the
database. Afterwards, the resource provider generation for each resource
provider involved in the allocation transaction has its generation
incremented. If another threads has concurrently updated either the
inventory or the usage of resources, this generation increment will
raise ConcurrentUpdateDetected, allowing callers to retry. This patch
does not yet implement this retry, which should be fine for the time
being considering that nothing yet calls POST /allocations.

Partially-Implements: blueprint generic-resource-pools

Co-Authored-By: Jay Pipes <jaypipes@gmail.com>
Co-Authored-By: Ed Leafe <ed@leafe.com>

Change-Id: Ic0bb48dcb1ab33b278f09dee8778fcb7a374caad