Updated openstack/openstack Project: openstack/heat-translator  68377c69bb63d639c57317e2d515af7a1b0784c2
Provide a specific stack_name when deploy a template use heat-translator

Currently, the stack name of a template is random when use
heat-translator online, and there is not any other feedback
information except success or failure, the result is not very convenient.
So it is necessary to provide a stack name when deploy a template, and
the stack name will be 'heat_' + filename + '_' + first part of a uuid if
no specific stack name provided.

Change-Id: Icd92835d4bdd8bbb7417210bccd648c0f2e4171b
Signed-off-by: shangxdy <shang.xiaodong@zte.com.cn>