Updated openstack/openstack Project: openstack/cinder  29e00a3c63e15cd87d60e6f7cdc661646e4cb816
Add supported driver checks to Zone Manager

This patch adds the supported driver checks to the
Fibre Channel Zone Manager.  This is the same mechanism
that is in Cinder volume drivers.

When a zone manager driver is marked as unsupported, the zone manager
won't allow it to work unless the enable_unsupported_driver=True is
added to the fc-zone-manager section in cinder.conf

Change-Id: If707b9005f60dc28eb99f46929fb700a2de0f755
Implements: blueprint zonemanager-supported-driver