Updated openstack/openstack Project: openstack/ceilometer  bf38c2c932ac1ee2b1c683949b254767b60948a1
Fix checking IP version when using IPv6.

This commit fixes an issue with ceilometer incorrectly determining the
IP version to use.  Previously, it checked whether the hostname was a
valid IPv6 address, but it should be resolving first to check whether
that is a valid IP address instead.

tests: Created test case that enables creating a UDPPublisher with
hostname for ipv4 and ipv6.

Closes-bug: #1616496

Change-Id: I586104fe69a03e64a0fb7572442b3175f880d80d
Co-authored-by: Joseph Richard <Joseph.Richard@windriver.com>