Updated openstack/openstack Project: openstack-infra/system-config  e611758bf824d2d6ba83ffeb10bceefcec985692
Add images profile for openstackci oscc clouds and add it to infracloud

Our openstackci projects all use Trusty for the mirror.
This image is uploaded manually by us, infra operators, every time
we spin up a new cloud.
Turns out we can automate this with this change from the launcher:

https://review.openstack.org/#/c/364397

With it, the launcher will download a trusty image from Ubuntu
site, put it on /tmp and upload it to the cloud.
This change will be a no-op, since the infracloud already has
a 'ubuntu-trusty' image but sets the pattern to follow to newer clouds
or in case we need to recreate an existing openstackci project.

Change-Id: Id1ef37ba46ac69eeb2da9a09bbc5f1ebd1d35afd