Updated openstack/openstack Project: openstack-infra/project-config  d4405275b3fb8c8fe67a8676d767c91b51b311fb
Migrate to gate-{name}-tox-db-{envlist}-{node}

This is part of our effort to migrate jobs from ubuntu-trusty to
ubuntu-xenial.

Note that some branch conditions need to be fixed:
Our goal is that ubuntu-trusty is used for stable/liberty and stable/mitaka
only (via global regex). And ubuntu-xenial is used for newer stable
branches and master.
Thus change a not on stable/liberty condition to a stable/mitaka
condition only for the trusty job.

Fix also a minor grammar issue in a comment in project.yaml.

Change-Id: I268c36611e34d378758a4094fa7d8df7189249b4
Signed-off-by: Paul Belanger <pabelanger@redhat.com>
Co-Authored-By: Andreas Jaeger <aj@suse.com>