Updated openstack/openstack Project: openstack/oslo.config  fcc344d9542e6dda5e6f4fb5c32919924a9c06bd
Warn user about enforce_type default change

We plan to change the enforce_type default to True.
https://review.openstack.org/#/c/328692

But this is going to break some project. This change
adds a deprecation warning to help them to track their broken
tests.

This should avoid people to change ton of tests file in all projects for
no real value. Specially for an option that in this end will be removed
from oslo.config when everybody have switched to the enforcing mode.

Related-Bug: #1517839

Change-Id: I438aeb766d663dbfe5dbd263fd166c25814204e8