Updated openstack/openstack Project: openstack/tempest  66070134bf583a4bc484233ced3cc172d88547ee
Remove the default size in volumes_client

The config value of volume size was passed to an instance of volumes_client
and it was used as the default value if not specifying a volume size from
tests side. However, the behavior made actual tests behaviors unreadable
when checking tests code.
In addition, this behavior is different from the other service clients.
So this patch removes the default size and makes tests specify the size
clearly.

Partially implements blueprint consistent-service-method-names

Change-Id: I2a0f5841fe82ab003562749a63bc4894225699c5