Updated openstack/openstack Project: openstack/heat  708157401354446bae7efed990ee209ca3f94cb6
Get rid of extra_rsrc_defn in templates

We have to hard-code knowledge of how to translate from template data to
ResourceDefinition objects, and we know what data we need in any of the
Template implementations to do so. So just get rid of the extra
indirection, non-locality of data, doubling up of definitions of valid
keys, and assumptions about the relationship between keys and
ResourceDefinition arguments and do everything directly for the condition
and external_id fields like we do for all of the other arguments.

Change-Id: I8e6381ed8ae7ffe4711653e473ab1412e4476be9