Updated openstack/openstack Project: openstack/cinder  60c47026f57f5b612d6ea9f844a0231b3adbc9ba
When only .pyc left, the extended API can not be used.

Currently, when only .pyc left, all extented apis, such as
resetting status of snapshot, can not be used.

This is because it only load .py files in api/extention.py.
__init__ and anything that's not .py are all skipped.

This change updates the extention api to load .pyc files
when .py files do not exist.

Change-Id: I80a3c29a50dcb3cb6bf3708dfeb2a3c75194065b
Closes-bug: #1616358