Updated openstack/openstack Project: openstack/cinder  2c623a8c96d8649f3a2963f85bf373782fcbc549
Fix driver interface checks for ZTE driver

This is part of the effort to gather accurate information
about Cinder drivers. The @interface.volumedriver decorator
should wrap all volume drivers. This is just a base class
for the ZteISCSIDriver. It does not implements all methods
of a volume driver and is not used as a driver.

Change-Id: Ib0c9fb5e41530ce51c24cbf0cd958e9ee89937ed