Updated openstack/openstack Project: openstack/tripleo-heat-templates  f86526e209d5040501b102e72a86cf31f9691f99
Ironic: update default enabled drivers

The goal is to have only production drivers available out of box.

Remove agent_ipmitool as it required additional configuration of
swift temporary URLs. Add pxe_ilo and pxe_drac as they work out of box.

This brings the default enabled drivers list closer to one of undercloud.
Note that pxe_ssh is not enabled as it's not a production driver.

Change-Id: Ic7d2190e2d39a795b024c2999e22ee47b3045cf1
Partially-Implements: blueprint ironic-integration