Updated openstack/openstack Project: openstack/heat  a5fb7635d222cc8a5338817224ee8434371a506d
Evaluate resource conditions in-place

This allows us to parse the conditions only once instead of reparsing them
all every time we encountered a resource with a condition. It also allows
us to get rid of some of the API surface area and reduces our reliance on
some of the rest.

Partial-Bug: #1618713

Change-Id: I3c2273722171b9c4cb13ef6588f7b522b6689b1c