Updated openstack/openstack Project: openstack/tempest  a07a423b3cf3348fa692dd9dae82f1d378a3f33a
Add skip logic for admin cred in test_auto_allocate_network

In case tests are being run with pre provisioned account and
admin cred are not defined then auto allocate network tests
which needs admin cred for tenant isolation should be skipped
instead of fail.

Those tests are not derived from admin base class(which is valid)
so we need to explicitly have the skip logic based on availability
of admin cred.

Change-Id: I3f0472665d7645986ee101390c8e406668a1567b