Updated openstack/openstack Project: openstack-infra/project-config  9d4cf90e85b6c673975ad48ad0e0a9d2f94cba43
Include both Trusty and Xenial gate-requirements-integration-dsvm

With the merge of Ia139c66e4e1daadae107b2fcdd0768d8d36bc984 we switched
to gating on ubuntu-xenial rather than ubuntu-trusty.  This uncovered a
small issue with the requirements gate and a larger issue for how we
correctly generate constraints during the transition from python3.4 to
python3.5.

This review adds back gate-requirements-integration-dsvm-ubuntu-trusty
so we can verify that we *can* correctly generate constraints for python
3.4 and python 3.5

Change-Id: I33e44578121c771e41d7068f9f020ecb87a9949f
Depends-On: I6acabf86933b7a7bce7baf44e6b512e7d0d2f6db
Related-Bug: 1620436