Updated openstack/openstack Project: openstack/swift  531a6d46e0e9eade7b305934964fa8e01bdb674e
Remove unnecessary open call and related vars in unit tests

Since the methods touching exisiting file is mocked, we don't have to
open the file. (if it is needed, we should close at first anyway)

And cleanup unecessary vars/imports that is used for making the file
path.

This patch is a followup on:
https://review.openstack.org/#/c/350471/

Change-Id: I8fbc0b5b9a01782b6da5a7dd674f52d4b566ca5c